feat(storybook): avoid refresh (#3841)

+++ apps/storybook/.storybook/preview.tsx
@@ -52,15 +52,21 @@ const ThemeChange = () => {
   return null;
 };
 
+const storeMap = new Map<string, ReturnType<typeof createStore>>();
+
 const withContextDecorator: Decorator = (Story, context) => {
   const { data: store } = useSWR(
     context.id,
     async () => {
+      if (storeMap.has(context.id)) {
+        return storeMap.get(context.id);
+      }
       localStorage.clear();
       const store = createStore();
       _setCurrentStore(store);
       await setup(store);
       await bootstrapPluginSystem(store);
+      storeMap.set(context.id, store);
       return store;
     },
     {
